# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vector Portal API
  version: 1.0.0
  description: |
    Document-workflow portal for multi-role approval chains.
    Manages events (document packages), their sections, and a
    role-based submit/approve/return pipeline with file attachments,
    inline comments, and a completed-document library.

    **Roles** (highest to lowest): ADMIN, PROTOCOL, DEPUTY, SUPERVISOR,
    SUPER_COLLABORATOR, COLLABORATOR, ANALYST.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by POST /api/auth/login (24 h expiry)
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]

    # ── Auth ──────────────────────────────────────────────────────────
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        isExternal:
          type: boolean
        entityName:
          type: string
          nullable: true
        mustChangePassword:
          type: boolean
    UserRole:
      type: string
      enum: [ADMIN, PROTOCOL, DEPUTY, SUPERVISOR, SUPER_COLLABORATOR, COLLABORATOR, ANALYST]
    DocumentSubmitterRole:
      type: string
      enum: [DEPUTY, SUPERVISOR, SUPER_COLLABORATOR]
    Language:
      type: string
      enum: [EN, FR, AR, ES, RU, ZH, PT, DE, KA]
    WorkflowType:
      type: string
      enum: [advanced, simple]
    EventStatus:
      type: string
      enum: [DRAFT, IN_PROGRESS, COMPLETED, ARCHIVED]

    # ── Events ────────────────────────────────────────────────────────
    EventSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        countryId:
          type: integer
        countryName:
          type: string
        countryCode:
          type: string
        documentSubmitterRole:
          $ref: '#/components/schemas/DocumentSubmitterRole'
        documentSubmitterId:
          type: integer
        documentSubmitterName:
          type: string
        deputyId:
          type: integer
          nullable: true
        supervisorId:
          type: integer
          nullable: true
        supervisorName:
          type: string
          nullable: true
        curatorRequired:
          type: boolean
        workflowType:
          $ref: '#/components/schemas/WorkflowType'
        language:
          $ref: '#/components/schemas/Language'
        deadlineDate:
          type: string
          format: date
        occasion:
          type: string
          nullable: true
        isActive:
          type: boolean
        endedAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/EventStatus'
        createdAt:
          type: string
          format: date-time
...
    SectionSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sortOrder:
          type: integer
        departmentIds:
          type: array
          items:
            type: integer
    SectionInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 500
        departmentIds:
          type: array
          items:
            type: integer

    # ── Workflow ───────────────────────────────────────────────────────
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        anchorId:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        userId:
          type: integer
        userName:
          type: string
        username:
          type: string

    # ── Files ─────────────────────────────────────────────────────────
    FileMetadata:
      type: object
      properties:
        id:
          type: integer
        original_name:
          type: string
        stored_name:
          type: string
        mime_type:
          type: string
        size:
          type: integer
        uploaded_by_id:
          type: integer
        uploaded_by_name:
          type: string
        created_at:
          type: string
          format: date-time

    # ── Library ───────────────────────────────────────────────────────
...
    LibraryDocument:
      type: object
      properties:
        eventId:
          type: integer
        title:
          type: string
        language:
          $ref: '#/components/schemas/Language'
        countryName:
          type: string
        endedAt:
          type: string
          format: date-time
          nullable: true
        sections:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              sortOrder:
                type: integer
              htmlContent:
                type: string
                nullable: true

    # ── Templates ─────────────────────────────────────────────────────
    Template:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        documentSubmitterRole:
          $ref: '#/components/schemas/DocumentSubmitterRole'
        curatorRequired:
          type: boolean
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdById:
          type: integer
          nullable: true
        createdByName:
          type: string
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionSummary'

    # ── Admin ─────────────────────────────────────────────────────────
...
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameEn:
          type: string
          nullable: true
        isExternal:
          type: boolean
...
paths:
...
  /api/users:
    get:
      tags: [users]
      summary: List all users (ADMIN only)
      responses:
        '200':
          description: Array of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
    post:
      tags: [users]
      summary: Create a user (ADMIN only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, username, email, password, role]
              properties:
                fullName:
                  type: string
                  maxLength: 200
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                role:
                  $ref: '#/components/schemas/UserRole'
                departmentId:
                  type: integer
                  nullable: true
                isExternal:
                  type: boolean
                entityName:
                  type: string
                  nullable: true
                countryIds:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  success:
                    type: boolean
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/departments:
    get:
      tags: [departments]
      summary: List all departments
      responses:
        '200':
          description: Array of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
    post:
      tags: [departments]
      summary: Create a department (ADMIN only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 200
                nameEn:
                  type: string
                  maxLength: 200
                  nullable: true
                isExternal:
                  type: boolean
      responses:
        '201':
          description: Department created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  success:
                    type: boolean
...
  /api/events:
    get:
      tags: [events]
      summary: List events visible to current user
      responses:
        '200':
          description: Array of event summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventSummary'
    post:
      tags: [events]
      summary: Create a new event
      description: Requires a role that can create events. ANALYST role is denied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, countryId, documentSubmitterRole, documentSubmitterId, deadlineDate, sections]
              properties:
                title:
                  type: string
                  maxLength: 500
                countryId:
                  type: integer
                documentSubmitterRole:
                  $ref: '#/components/schemas/DocumentSubmitterRole'
                documentSubmitterId:
                  type: integer
                deputyId:
                  type: integer
                  nullable: true
                supervisorId:
                  type: integer
                  nullable: true
                curatorRequired:
                  type: boolean
                workflowType:
                  $ref: '#/components/schemas/WorkflowType'
                language:
                  $ref: '#/components/schemas/Language'
                deadlineDate:
                  type: string
                  format: date
                occasion:
                  type: string
                  nullable: true
                sections:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items:
                    $ref: '#/components/schemas/SectionInput'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  success:
                    type: boolean
        '403':
          description: Not authorized
        '422':
          description: Validation error (e.g. invalid document submitter)
...
  /api/sections:
    get:
      tags: [sections]
      summary: List sections for an event
      parameters:
        - name: event_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Array of sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SectionSummary'
...
  /api/workflow/comments:
    get:
      tags: [comments]
      summary: List comments for a section
      parameters:
        - name: event_id
          in: query
          required: true
          schema:
            type: integer
        - name: section_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Array of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
    post:
      tags: [comments]
      summary: Add a comment on a section
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, sectionId, content]
              properties:
                eventId:
                  type: integer
                sectionId:
                  type: integer
                parentId:
                  type: integer
                  nullable: true
                anchorId:
                  type: string
                  nullable: true
                  maxLength: 200
                content:
                  type: string
                  maxLength: 10000
                htmlContent:
                  type: string
                  nullable: true
                  description: Auto-saves editor HTML to persist anchor
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  success:
                    type: boolean
...
  /api/library/{eventId}/document:
    get:
      tags: [library]
      summary: Get full document with all section HTML
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Full document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryDocument'
  /api/library/{eventId}/files:
    get:
      tags: [library]
      summary: List all files across sections for a library event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Array of file metadata with section titles
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/FileMetadata'
                    - type: object
                      properties:
                        section_title:
                          type: string
  /api/library/{eventId}/reopen:
    post:
      tags: [library]
      summary: Reopen a completed event for amendment (DS only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Event reopened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Event is not COMPLETED
        '403':
          description: Only the Document Submitter can reopen

  # ── Templates ─────────────────────────────────────────────────────
  /api/templates:
    get:
      tags: [templates]
      summary: List event templates (defaults + own)
      responses:
        '200':
          description: Array of templates with sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
    post:
      tags: [templates]
      summary: Create an event template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 300
                documentSubmitterRole:
                  $ref: '#/components/schemas/DocumentSubmitterRole'
                curatorRequired:
                  type: boolean
                sections:
                  type: array
                  maxItems: 100
                  items:
                    $ref: '#/components/schemas/SectionInput'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  success:
                    type: boolean
...